+++ test/brando/images/utils_test.exs
@@ -1,6 +1,11 @@
 defmodule Brando.Images.UtilsTest do
   use ExUnit.Case
+  use Brando.ConnCase
+  use Plug.Test
+  use RouterHelper
+
   import Brando.Images.Utils
+
   alias Brando.Images.Upload
   alias Brando.Factory
 
@@ -46,5 +51,33 @@ defmodule Brando.Images.UtilsTest do
     assert media_path == Brando.config(:media_path)
     assert media_path(nil) == Brando.config(:media_path)
     assert media_path("images") == Path.join(Brando.config(:media_path), "images")
+  end
+
+  test "put_size_cfg" do
+    user      = Factory.insert(:user)
+    category  = Factory.insert(:image_category, creator: user)
+    series    = Factory.insert(:image_series, creator: user, image_category: category)
+
+    up_params = Factory.build(:plug_upload)
+
+    :post
+    |> call("/admin/images/series/#{series.id}/upload", %{"id" => series.id, "image" => up_params})
+    |> with_user(user)
+    |> as_json
+    |> send_request
+
+    put_size_cfg(series, "medium", %{"portrait" => %{"size" => "500", "quality" => 1},
+                                     "landscape" => %{"size" => "50", "quality" => 1}})
+
+    series = Brando.repo.get(Brando.ImageSeries, series.id)
+
+    assert series.cfg.sizes == %{
+      "medium" => %{
+        "landscape" => %{"quality" => 1, "size" => "50"},
+        "portrait" => %{"quality" => 1, "size" => "500"}
+      },
+      "small" => %{"quality" => 1, "size" => "300"},
+      "thumb" => %{"crop" => true, "quality" => 1, "size" => "150x150"}
+    }
   end
 end