+++ test/support/factory.ex
@@ -5,8 +5,9 @@ defmodule Brando.Factory do
   alias Brando.{ImageCategory, ImageSeries, User}
 
   @sizes %{
-    "small" =>  %{"size" => "300", "quality" => 100},
-    "thumb" =>  %{"size" => "150x150", "quality" => 100, "crop" => true}
+    "small" =>  %{"size" => "300", "quality" => 1},
+    "medium" => %{"portrait" => %{"size" => "300", "quality" => 1}, "landscape" => %{"size" => "250", "quality" => 1}},
+    "thumb" =>  %{"size" => "150x150", "quality" => 1, "crop" => true}
   }
 
   def user_factory do
@@ -76,12 +77,12 @@ defmodule Brando.Factory do
       random_filename: false,
       size_limit: 10_240_000,
       sizes: %{
-        "small" =>  %{"size" => "300", "quality" => 100},
-        "medium" => %{"size" => "500", "quality" => 100},
-        "large" =>  %{"size" => "700", "quality" => 100},
+        "small"  => %{"size" => "300", "quality" => 100},
+        "medium" => %{"portrait" => %{"size" => "300", "quality" => 1}, "landscape" => %{"size" => "250", "quality" => 1}},
+        "large"  => %{"size" => "700", "quality" => 100},
         "xlarge" => %{"size" => "900", "quality" => 100},
-        "thumb" =>  %{"size" => "150x150", "quality" => 100, "crop" => true},
-        "micro" =>  %{"size" => "25x25", "quality" => 100, "crop" => true}
+        "thumb"  => %{"size" => "150x150", "quality" => 100, "crop" => true},
+        "micro"  => %{"size" => "25x25", "quality" => 100, "crop" => true}
       }
     }
   end